refactor(llc): move the event handler and state mutations into the channel directory
