+++ packages/stream_chat/lib/src/client/channel/channel_event_handler.dart
@@ -1,4 +1,4 @@
-import '../../stream_chat.dart';
+import '../../../stream_chat.dart';
 import 'channel_state_mutations.dart';
 
 /// Routes channel events to the matching [ChannelStateMutations] methods.
