+++ packages/stream_chat/test/src/client/channel/channel_state_mutations_test.dart
@@ -1,12 +1,12 @@
 // ignore_for_file: cascade_invocations
 
 import 'package:mocktail/mocktail.dart';
-import 'package:stream_chat/src/client/channel_state_mutations.dart';
+import 'package:stream_chat/src/client/channel/channel_state_mutations.dart';
 import 'package:stream_chat/stream_chat.dart';
 import 'package:test/test.dart';
 
-import '../fakes.dart';
-import '../mocks.dart';
+import '../../fakes.dart';
+import '../../mocks.dart';
 
 class MockChannelClientState extends Mock implements ChannelClientState {}
